refactor: 타입 경로를 상대 경로에서 절대 경로로 변경
