+++ src/components/LoginForm/LoginForm.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 
-import type { LoginFormProps } from '../../types/login';
+import type { LoginFormProps } from '@/types/login';
+
 import Button from '../atoms/Button/Button';
 import Input from '../atoms/Input/Input';
 import PasswordInput from '../atoms/Input/PasswordInput';
