+++ src/components/Modals/AgreementModal.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 
-import type { AgreementModalProps } from '../../types/modal';
+import type { AgreementModalProps } from '@/types/modal';
+
 import Modal from '../Modal/Modal';
 import ModalCloseButton from '../Modal/ModalCloseButton';
 import ModalHeader from '../Modal/ModalHeader';
