+++ src/components/SignupForm/SignupForm.tsx
@@ -1,6 +1,7 @@
+import type { SignupValues } from '@/types/signupForm.types';
+import { signupInitialForm } from '@/types/signupForm.types';
+
 import { useSignupForm } from '../../hooks/useSignupForm';
-import type { SignupValues } from '../../types/signupForm.types';
-import { signupInitialForm } from '../../types/signupForm.types';
 import Button from '../atoms/Button/Button';
 import Input from '../atoms/Input/Input';
 import PasswordInputField from '../atoms/Input/PasswordInputField';
